changed disc cutoff to 12.5

+++ Model/VoucherCost.cs
@@ -4,7 +4,7 @@
     {
         public string? VoucherValue { get; set; }
         public string? VoucherCosting { get; set; }
-        public int? DiscountRate { get; set; }
+        public decimal? DiscountRate { get; set; }
         public int? PageNum { get; set; }
 
     }
